Moved map code into the map screen class, also started tidying up some naming convention things.

+++ TurnBased/src/inf1/oop/turnbased/screen/Screen.java
@@ -1,6 +1,6 @@
 package inf1.oop.turnbased.screen;
 
 public abstract class Screen {
-	public abstract void draw(int dt);
-	public abstract void update(int dt);
+	public abstract void draw(float dt);
+	public abstract void update(float dt);
 }
